fix(review): move _check_script_version_health from _doctor_config to _doctor_fleet
